+++ packages/js/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @trainheroic-unofficial/js
 
+## 0.6.1
+
+### Patch Changes
+
+- @trainheroic-unofficial/dto@0.6.1
+
 ## 0.6.0
 
 ### Patch Changes